词法解析器：自动机 Demo 调通

+++ include/lex_type/lex_string.h
@@ -10,7 +10,7 @@
  */
 struct LEX_STRING {
     char *str{};
-    size_t length{};
+    int length{};
 };
 
 #endif //LEX_STRING_H
